#41 delete user prompt

+++ songify-frontend/src/components/admin/users.list.component.jsx
@@ -1,37 +1,43 @@
-import React, {Component} from 'react';
+import React, { Component } from 'react';
 import UserService from '../../services/user.service.js';
 
-export default class Users extends Component{
+export default class Users extends Component {
 
-    constructor(props){
+    constructor(props) {
         super(props)
         this.state = {
-            users:[]
+            users: []
         }
     }
 
-    componentDidMount(){
-        UserService.getUsers().then((response)=>{
-            this.setState({ users: response.data})
+    componentDidMount() {
+        UserService.getUsers().then((response) => {
+            this.setState({ users: response.data })
         });
     }
 
-    editUser(id){
+    editUser(id) {
         this.props.history.push(`/update-user/${id}`);
     }
 
-    deleteUser(id){
-        UserService.deleteUser(id).then((response)=>{
-            this.setState({ users: this.state.users.filter((user) => user.id !== id)}); 
-        });
+    deleteUser(id) {
+        var r = window.confirm("Are you sure you want to delete this user?");
+        if (r == true) {
+            UserService.deleteUser(id).then((response) => {
+                this.setState({ users: this.state.users.filter((user) => user.id !== id) });
+            });
+        } else {
+            window.location.reload();
+        }
+       
     }
 
-    render (){
+    render() {
         return (
             <div>
-                <h1 className = "text-center"> Users List</h1>
+                <h1 className="text-center"> Users List</h1>
 
-                 <table className = "table table-striped">
+                <table className="table table-striped">
                     <thead>
                         <tr>
                             <td>ID</td>
@@ -47,18 +53,18 @@ export default class Users extends Component{
                         {
                             this.state.users.map(
                                 user =>
-                                <tr key={user.id}>
-                                    <td>{user.id}</td>
-                                    <td>{user.email}</td>
-                                    <td>{user.username}</td>
-                                    <td>{user.password}</td>
-                                    <td>{user.role.id}</td>
-                                    <td>{user.role.name}</td>
-                                    <td>              
-                                        <button type="button" className="btn btn-primary" onClick={() => this.editUser(user.id)} >Edit</button>
-                                        <button type="button" className="btn btn-danger" onClick={() => this.deleteUser(user.id)} >Delete</button>
-                                    </td>
-                                </tr>
+                                    <tr key={user.id}>
+                                        <td>{user.id}</td>
+                                        <td>{user.email}</td>
+                                        <td>{user.username}</td>
+                                        <td>{user.password}</td>
+                                        <td>{user.role.id}</td>
+                                        <td>{user.role.name}</td>
+                                        <td>
+                                            <button type="button" className="btn btn-primary" onClick={() => this.editUser(user.id)} >Edit</button>
+                                            <button type="button" className="btn btn-danger" onClick={() => this.deleteUser(user.id)} >Delete</button>
+                                        </td>
+                                    </tr>
                             )
                         }
                     </tbody>
